dapr-quickstart/order-porcessor/app.py
import logging

from dapr.ext.fastapi import DaprApp
from fastapi import FastAPI
from pydantic import BaseModel

logging.basicConfig(
    level=logging.INFO,
    format="%(asctime)s.%(msecs)03d:%(levelname)s:%(name)s:%(message)s",
    datefmt="%Y-%m-%d %H:%M:%S",
)

app = FastAPI()
dapr_app = DaprApp(app)


class CloudEvent(BaseModel):
    datacontenttype: str
    source: str
    topic: str
    pubsubname: str
    data: dict
    id: str
    specversion: str
    tracestate: str
    type: str
    traceid: str


# Dapr subscription routes orders topic to this route
@dapr_app.subscribe(pubsub="orderpubsub", topic="orders")
def orders_subscriber(event: CloudEvent):
    logging.debug(event)
    logging.info(f"Subscriber received : {event.data['order_id']}")
    return {"success": True}
