dapr-quickstart/order-porcessor/requirements.txt
fastapi
dapr-ext-fastapi>=1.10.0,<2.0.0
cloudevents==1.10.0
uvicorn